refactor: thermos CQ round 3 final fixes
