+++ src/mergecraft/mcp/server.py
@@ -375,9 +375,6 @@ def create_mcp_app(
             return JSONResponse({"error": str(exc)}, status_code=401)
         except (ValueError, json.JSONDecodeError) as exc:
             return JSONResponse({"error": str(exc)}, status_code=400)
-        except Exception:
-            logger.exception("unexpected webhook ingress failure for provider={}", provider)
-            raise
         return JSONResponse(
             {
                 "result_id": result.result_id,
